Add back FakeImageLoader but limit it to no-op
